requirements.txt
aiosmtplib==2.0.0
celery==5.2.7
fastapi==0.85.1
fastapi-mail==1.2.0
jinja2==3.1.2
orjson==3.8.0
pika==1.3.1
pydantic==1.10.2
python-dotenv==0.21.0
uvicorn==0.17.6
